% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choropleths.R
\name{choropleths}
\alias{choropleths}
\title{Create Choropleths Object from Input Data}
\usage{
choropleths(df, which_map, id_data_index = c(1, 2))
}
\arguments{
\item{df}{Data you want to plot choropleth maps with. Should at least have two columns: region ID and data. Region ID column should be either DMA code or state code. The class type of region ID column should be character.}

\item{which_map}{Which map to plot. Either "DMA" or "state".}

\item{id_data_index}{Column index of region ID column and data column. Defaults to c(1, 2) when the first column is region ID and the second column is data.}
}
\description{
This function allows you to create \code{choropleths} object from input data indexed by region ID, which you can use to create choropleths maps.
}
\examples{
library(choropleths)

choro <- choropleths(sample_data_state, "state")

draw_map(choro)
draw_legend(choro)
}

