% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choropleths.R
\name{draw_legend.choropleths}
\alias{draw_legend.choropleths}
\title{Draw Choropleth Map Legend from Choropleths Object}
\usage{
\method{draw_legend}{choropleths}(obj, colors = NULL, domain = NULL,
  digits = 3, percent = FALSE)
}
\arguments{
\item{obj}{\code{choropleths} object}

\item{colors}{A vector of color codes for color scales.If not specified, defaults to \code{c("#1A9850", "#66BD63", "#A6D96A", "#D9EF8B", "#FFFFBF", "#FEE08B", "#FDAE61", "#F46D43", "#D73027")}, which changes from Green, Yellow and to Red as the value of input data increases.}

\item{domain}{A vector of length 2 for the lower bound and upper bound of input data for color scales. Any values outside of domain will be treated as either lower or upper bound. If not specified, the lower bound and upper bound of input data will be used as domain.}

\item{digits}{Maximum digits to print after decimal point. Defaults to 3.}

\item{percent}{wheather to print the data as percentage in the legend.}
}
\description{
This function allows you to draw choropleth map legend from \code{choropleths} object. The choropleth map legend will be drawn with ggplot2. You can configure colors to scale and domain of input data. You can also configure the format of data in the legend.
}
\examples{
library(choropleths)

choro <- choropleths(sample_data_state, "state")

draw_map(choro)
draw_legend(choro)
}

